requirements.txt
Django==2.1
django-cleanup==5.0.0
django-crispy-forms==1.7.2
django-hitcount==1.3.2
easy-thumbnails==2.7
Pillow
python-decouple==3.3
pytz==2020.1
sorl-thumbnail==12.6.3
djangorestframework==3.11.0
pyaxmlparser==0.3.24
django-extensions==3.1.0
